improve: Only load async workers if deps available
