+++ lib/vero.rb
@@ -5,6 +5,9 @@ require "json"
 require "rest-client"
 require "zeitwerk"
 
+module Vero
+end
+
 loader = Zeitwerk::Loader.for_gem
 loader.inflector.inflect(
   "api_context" => "APIContext",
@@ -19,6 +22,12 @@ loader.inflector.inflect(
   "unsubscribe_api" => "UnsubscribeAPI"
 )
 loader.ignore("#{__dir__}/generators")
+
+loader.push_dir("#{__dir__}/vero/workers", namespace: Vero)
+loader.ignore("#{__dir__}/vero/workers/resque_worker") unless defined?(Resque)
+loader.ignore("#{__dir__}/vero/workers/sidekiq_worker") unless defined?(::Sidekiq)
+loader.ignore("#{__dir__}/vero/workers/sucker_punch_worker") unless defined?(SuckerPunch)
+
 loader.setup
 
 require "vero/railtie" if defined?(Rails)
